requirements.txt
kivy==2.3.1
pandas>=1.3.0
numpy>=1.21.0
akshare>=1.8.0
matplotlib>=3.4.0
pillow>=8.3.0 